refactor: 투표 관련 파일 분류
