+++ src/app/(route)/votes/[voteId]/_component/ManagementButton.tsx
@@ -3,7 +3,7 @@
 import { useRef, useState } from 'react'
 import { useRouter } from 'next/navigation'
 import Image from 'next/image'
-import { deleteVote } from '@/app/_hook/api/useDeleteVote'
+import { deleteVote } from '@/app/_hook/api/votes/useDeleteVote'
 import useOutsideClick from '@/app/_hook/common/useOutsideClick'
 
 export default function ManagementButton({ voteId }: { voteId: number }) {
