+++ src/app/(route)/votes/[voteId]/_component/VoteInfo.tsx
@@ -1,9 +1,9 @@
 'use client'
 
 import React, { useCallback, useState } from 'react'
-import { voteItem } from '@/app/_hook/api/useParticipationVote'
+import { voteItem } from '@/app/_hook/api/votes/useParticipationVote'
 import { ItemInfoType, VoteInfoType } from '@/app/_types/detailVote.type'
-import { reVote } from '@/app/_hook/api/useReParticipation'
+import { reVote } from '@/app/_hook/api/votes/useReParticipation'
 import ProgressBar from './ProgressBar'
 import VoteItem from './VoteItem'
