+++ src/app/_hook/api/votes/useVoteDetail.ts
